Add files via upload
Hello, Bill.  This file I just added is generated directly from our PMS (property management software). It is does not give an option to generate as a CSV so I downloaded as a standard latest version XCEL and then used XCEL to convert to .csv8 (comma delimited). I changed nothing but one thing just for your info.  I edited some blocks, just for visual reference to be highlighed yellow.  IMHO, this is the only info we need.  Those specific data points in the Bldg/Unit column.  Those are our unique identifiers for our vacant units.  ** Note the other version I could generate this form in is XML  which comes out as a web address and looks structurally similar to a json file, but I'm not familiar with it so I wasn't sure if we could use it or not as it is likely pw protected, however I'm including it here in the note, as well as a rich text of the raw date if the link doesn't work: https://mutualmanagement.onesite.realpage.com/SDESvc/BL/HomePageListBL.asmx/SaveAs?FileID=1354206&view=1&type=%20Report