Add files via upload
Bill, added a new CSV from our PMS on Availability and it includes the units we disdussed Sunday, the occupied but with notice given to vacate ( Section "NTV" ) and the format is the same as the others we started working already.  So we're good with Plan A.   

However, another little task we'll have to manage is file formatting.  The PMS software only gives the option to download Xcel formatting.  I'm saving to CSV manually on my harddrive.  We'll cross that bridge when we get to it.